Fix: add default for the new custom template
The addition of that template variable to specify a custom template broke PDF generation for anyone who already had this package in use, since we didn't expect to have specify a template. This PR seeks to set a sensible default for that variable based on the previously recommend location of the default invoices template.